migration events request

+++ src/main/java/fr/m2i/fil_rouge_spring/models/Collaborator.java
@@ -20,7 +20,7 @@ public class Collaborator {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Integer _id;
+    private Long _id;
 
     @Basic
     @Column(name = "user_id")
@@ -56,11 +56,11 @@ public class Collaborator {
         this.share = share;
     }
 
-    public Integer get_id() {
+    public Long get_id() {
         return _id;
     }
 
-    public void set_id(Integer _id) {
+    public void set_id(Long _id) {
         this._id = _id;
     }
